created peak frequency for the data
